+++ src/Braintree/CreditCard.cs
@@ -87,6 +87,7 @@ namespace Braintree
         public static readonly CreditCardCardType AMEX = new CreditCardCardType("American Express");
         public static readonly CreditCardCardType CARTE_BLANCHE = new CreditCardCardType("Carte Blanche");
         public static readonly CreditCardCardType CHINA_UNION_PAY = new CreditCardCardType("China UnionPay");
+        public static readonly CreditCardCardType UNION_PAY = new CreditCardCardType("UnionPay");
         public static readonly CreditCardCardType DINERS_CLUB_INTERNATIONAL = new CreditCardCardType("Diners Club");
         public static readonly CreditCardCardType DISCOVER = new CreditCardCardType("Discover");
         public static readonly CreditCardCardType JCB = new CreditCardCardType("JCB");